+++ izanami-server/conf/application.conf
@@ -221,7 +221,7 @@ izanami {
     throughput = 1
   }
   oauth2 {
-    enabled = false
+    enabled = true
     authorizeUrl = "http://localhost:8980/auth/realms/master/protocol/openid-connect/auth"
     tokenUrl = 	"http://localhost:8980/auth/realms/master/protocol/openid-connect/token"
     userInfoUrl = "http://localhost:8980/auth/realms/master/protocol/openid-connect/userinfo"
